Fixes meta tags on route activated pages
