+++ schema/page/intro.ts
@@ -22,7 +22,8 @@ import {
     TitleBodyActions,
 } from "../text";
 import { PropsOf } from "../util";
-import { Page, SanityPage, metaTagsField } from "./common";
+import { Page, SanityPage } from "./common";
+import { metaTagsField } from "./meta-tags";
 
 export interface SanityIntroPage extends SanityPage {
     introSection: SanityTitleBodyActions;
