+++ schema/page/solution.ts
@@ -32,7 +32,8 @@ import {
     TitleAndBody,
 } from "../text";
 import { PropsOf } from "../util";
-import { metaTagsField, Page, SanityPage } from "./common";
+import { Page, SanityPage } from "./common";
+import { metaTagsField } from "./meta-tags";
 
 const sections = {
     intro: { id: "introSection", title: "Intro" },
