+++ schema/webinar.ts
@@ -19,7 +19,7 @@ import { personSchemaName } from "./person";
 import { SanityDataset } from "./sanity-core";
 import { PropsOf } from "./util";
 import { EventBase, SanityEventBase } from "./event-base";
-import { metaTagsField } from "./page/common";
+import { metaTagsField } from "./page/meta-tags";
 
 export interface SanityWebinar extends SanityEventBase {
     datetime: string;
