+++ website/src/page/support-page/support-page.component.ts
@@ -36,7 +36,8 @@ export class SupportPageComponent implements OnInit {
             tap((page) => {
                 if (page) {
                     this.title.setTitle(`${page.title} - TypeDB`);
-                    this.metaTags.register(page.metaTags, this.destroyRef);
+                    const { unregister } = this.metaTags.register(page.metaTags);
+                    this.destroyRef.onDestroy(unregister);
                     this.analytics.hubspot.trackPageView();
                     setTimeout(() => {
                         this.idleMonitor.fireManualMyAppReadyEvent();
