+++ website/src/page/white-papers-page/white-papers-page.component.ts
@@ -38,7 +38,8 @@ export class WhitePapersPageComponent implements OnInit {
             if (sanityWhitePapersPage) {
                 this.page = new WhitePapersPage(sanityWhitePapersPage, data);
                 this._title.setTitle(`${this.page.title} - TypeDB`);
-                this.metaTags.register(this.page.metaTags, this.destroyRef);
+                const { unregister } = this.metaTags.register(this.page.metaTags);
+                this.destroyRef.onDestroy(unregister);
                 this._analytics.hubspot.trackPageView();
                 setTimeout(() => {
                     this._idleMonitor.fireManualMyAppReadyEvent();
